First draft of SummaryDds
